# user-rest-service-1.0.2/generated-code/python/SwaggerPetstore-python/SwaggerPetstore/userwithresponseentity_api.py
#!/usr/bin/env python
# coding: utf-8

"""
UserwithresponseentityApi.py
Copyright 2015 Reverb Technologies, Inc.

   Licensed under the Apache License, Version 2.0 (the "License");
   you may not use this file except in compliance with the License.
   You may obtain a copy of the License at

       http://www.apache.org/licenses/LICENSE-2.0

   Unless required by applicable law or agreed to in writing, software
   distributed under the License is distributed on an "AS IS" BASIS,
   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
   See the License for the specific language governing permissions and
   limitations under the License.

NOTE: This class is auto generated by the swagger code generator program. Do not edit the class manually.
"""
import sys
import os
import urllib

from models import *



class UserwithresponseentityApi(object):

    def __init__(self, apiClient):
      self.apiClient = apiClient

    
    
    def getUserResponseEntity(self, **kwargs):
        """getUserResponseEntity

        Args:
            
            id, long: id (required)
            
            
        
        Returns: object
        """

        allParams = ['id']

        params = locals()
        for (key, val) in params['kwargs'].iteritems():
            if key not in allParams:
                raise TypeError("Got an unexpected keyword argument '%s' to method getUserResponseEntity" % key)
            params[key] = val
        del params['kwargs']

        resourcePath = '/user_with_responseentity'
        resourcePath = resourcePath.replace('{format}', 'json')
        method = 'GET'

        queryParams = {}
        headerParams = {}
        formParams = {}
        files = {}
        bodyParam = None

        headerParams['Accept'] = '*/*'
        headerParams['Content-Type'] = 'application/json,'

        
        if ('id' in params):
            queryParams['id'] = self.apiClient.toPathValue(params['id'])
        

        

        

        

        

        postData = (formParams if formParams else bodyParam)

        response = self.apiClient.callAPI(resourcePath, method, queryParams,
                                          postData, headerParams, files=files)

        
        if not response:
            return None

        responseObject = self.apiClient.deserialize(response, 'object')
        return responseObject
        
        
        
    